Add CaptureManager and UI components for screen capture functionality
- Implemented CaptureManager to handle screenshot, video, and GIF recording.
- Added MenuBarContentView for menu bar interactions, including shortcuts for capture actions.
- Created ProcessingIndicatorWindow to show processing status during captures.
- Integrated onboarding and guide windows for user assistance.
- Enhanced mouse click monitoring for video and GIF captures.
