+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Fixed
+- Processing indicator after recording stops now reliably appears and animates. Replaced the custom spinner (which failed to animate in borderless `NSPanel` + `NSHostingView`) with the system `ProgressView`. Added a spring fade-in entrance and a smooth window-level fade-in via `NSAnimationContext`.
+- Eliminated the blank gap between the processing indicator dismissing and the trimmer window appearing by moving `dismissProcessingIndicator()` to after `showTrimmer`/`showGifTrimmer` is called.
+
 ## v1.3.4 - 2026-04-05
 
 ### Improved
